Added crude implementation of delete function
Added a crude function to remove words from the trie. It does not check
that the word it is trying to delete exists, which can cause it to
derefence a NULL pointer. In addition, it will free the head of the trie
if there are no other words in it. These should be fixed later.