delete function now uses a pointer to head
The function uses a pointer to head so that it can manipulate the value
of head e.g. setting head to NULL within the function instead of relying
on the calling function. This also frees up the return values and means
that we can set the head of the entire trie to NULL to signal that the
memory was freed so that other functions can reallocate memory if
needed.